lambda_looker/package/requirements.txt
google-cloud-bigquery==3.17.2
requests==2.31.0
google-auth==2.24.0
python-dotenv==1.0.0